Moved out profile member data out of profiles and into its own subfolder for better management.
