+++ hypixel_api_lib/member/__init__.py
@@ -1,0 +1,2 @@
+from .ProfileMember import SkyBlockProfileMember
+from .ProfileMember import DeletionNotice
